@@ -1,86 +1,39 @@
 name: amparo_app
 description: Aplicativo amparo
-
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
+publish_to: 'none'
+
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  http: ^0.13.3
   cupertino_icons: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
   assets:
-     - assets/fonts/
-     - assets/images/logo.png
-     - assets/images/frame.png
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
+    - assets/certificates/
+    - assets/fonts/
+    - assets/images/logo.png
+    - assets/images/frame.png
   fonts:
-     - family: Archivo
-       fonts:
-         - asset: assets/fonts/Archivo-Light.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-
-     - family: SF Pro
-       fonts:
-         - asset: assets/fonts/SFProDisplay-Bold.ttf
-  #         weight: 700
-  #
-     - family: Audiowide
-       fonts:
-          - asset: assets/fonts/Audiowide-Regular.ttf
-#
-     - family: SF Pro Medium
-       fonts:
-          - asset: assets/fonts/SFProDisplay-Medium.ttf
-
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
+    - family: Archivo
+      fonts:
+        - asset: assets/fonts/Archivo-Light.ttf
+    - family: SF Pro
+      fonts:
+        - asset: assets/fonts/SFProDisplay-Bold.ttf
+    - family: Audiowide
+      fonts:
+        - asset: assets/fonts/Audiowide-Regular.ttf
+    - family: SF Pro Medium
+      fonts:
+        - asset: assets/fonts/SFProDisplay-Medium.ttf
